+++ tests/unit_tests/Commands/Solr/DisableCommandTest.php
@@ -32,18 +32,11 @@ class DisableCommandTest extends CommandTestCase
             ->disableOriginalConstructor()
             ->getMock();
         $this->solr->expects($this->once())
-            ->method('disable');
-        $this->site->expects($this->once())
-            ->method('converge')
+            ->method('disable')
             ->willReturn($workflow);
         $this->site->method('getSolr')->willReturn($this->solr);
 
         $this->logger->expects($this->at(0))
-            ->method('log')->with(
-                $this->equalTo('notice'),
-                $this->equalTo('Solr disabled. Converging bindings.')
-            );
-        $this->logger->expects($this->at(1))
             ->method('log')->with(
                 $this->equalTo('notice'),
                 $this->equalTo('successful workflow')