+++ CHANGELOG.md
@@ -8,6 +8,17 @@ All notable changes to this project will be documented in this file. This projec
 - New option `no-db` added to `Environments::create` which will skip the duplication of the database from the source environment. (#2050)
 - New option `no-files` added to `Environments::create` which will skip the duplication of files from the source environment. (#2050)
 
+### Changed
+- Return type of `Redis::disable()` has changed to a `Workflow` model. (#2XXX)
+- Return type of `Redis::enable()` has changed to a `Workflow` model. (#2XXX)
+- Return type of `Solr::disable()` has changed to a `Workflow` model. (#2XXX)
+- Return type of `Solr::enable()` has changed to a `Workflow` model. (#2XXX)
+- Return type of `Environment::disableHttpsCertificate()` has changed to a `Workflow` model. (#2XXX)
+
+### Removed
+- Removed obsolete `Environment::convergeBindings()` method. (#2XXX)
+- Removed obsolete `Site::converge()` method. (#2XXX)
+
 ## 2.3.0 - 2020-01-10
 ### Added
 - New const `TERMINUS_HOST_CERT` added to permit use of SSL certificates while making requests. (#2024)